Working around long path name issues on Windows by renaming PreserveCo... It was weird to have been nested, anyway.
